require *full* match against database items to return a GIN match
We are getting too many bad partial matches, e.g. "hot sauce" -> "apple sauce".
If it doesn't say "apple" AND "sauce", it's not apple sauce. We have the whole
idea of dependent and independent name variants for exactly this reason. If
something isn't specified as being identifiable by a subset of its tokens, we
shouldn't be identifying it.